Extract candidateAccountForm component
